Cleanups in buffer_core.cpp.
This does a few things:
1.  Removes commented out code.  This code has been commented
out for a decade, so I think it is safe to say nobody is
using it.
2.  Uses size_t everywhere we are calling vector.size()
3.  Ensures all member variables are suffixed with _
4.  Removes asserts and replaces them with exceptions.
5.  Gets rid of redundant branches for error cases and
uses just one switch statement for all possible return values.

+++ tf2/include/tf2/transform_storage.h
@@ -37,7 +37,7 @@
 
 namespace tf2
 {
-typedef uint32_t CompactFrameID;
+typedef size_t CompactFrameID;
 
 /** \brief Storage for transforms and their parent */
 class TransformStorage
